requirements/base.txt
Django==1.6.5
django-extensions==1.3.8
six==1.7.3
wsgiref==0.1.2
simplejson==3.5.3
requests==2.3.0
djangorestframework==2.3.14
django-filter==0.7
South==0.8.4
